+++ sources/packages/web/src/components/common/ManageProgramAndOfferingSummary.vue
@@ -1,7 +1,12 @@
 <template>
+  <ProgramDetailHeader :educationProgram="educationProgram" />
   <v-card>
     <v-container>
-      <ProgramDetails :programId="programId" :locationId="locationId" />
+      <ProgramDetails
+        :programId="programId"
+        :locationId="locationId"
+        :educationProgram="educationProgram"
+      />
       <hr class="horizontal-divider" />
       <OfferingSummary :programId="programId" :locationId="locationId" />
     </v-container>
@@ -9,11 +14,14 @@
 </template>
 
 <script lang="ts">
-import ProgramDetails from "@/components/common/ProgramDetails.vue";
+import ProgramDetails from "@/components/common/ProgramDetail.vue";
 import OfferingSummary from "@/components/common/OfferingSummary.vue";
+import ProgramDetailHeader from "@/components/common/ProgramDetailHeader.vue";
+
+import { EducationProgramData } from "@/types";
 
 export default {
-  components: { ProgramDetails, OfferingSummary },
+  components: { ProgramDetails, OfferingSummary, ProgramDetailHeader },
   props: {
     programId: {
       type: Number,
@@ -22,6 +30,11 @@ export default {
     locationId: {
       type: Number,
       required: true,
+    },
+    educationProgram: {
+      type: Object,
+      required: true,
+      default: {} as EducationProgramData,
     },
   },
 };
