+++ sources/packages/web/src/views/aest/institution/Locations.vue
@@ -1,9 +1,7 @@
 <template>
   <v-card class="mt-4">
     <div class="mx-5 py-4">
-      <LocationSummary
-        :institutionId="institutionId"
-      />
+      <LocationSummary :institutionId="institutionId" />
     </div>
   </v-card>
 </template>
@@ -18,6 +16,6 @@ export default {
       type: Number,
       required: true,
     },
-  }
+  },
 };
 </script>
