+++ sources/packages/web/src/views/aest/institution/ProgramDetails.vue
@@ -11,6 +11,7 @@
     <ManageProgramAndOfferingSummary
       :programId="programId"
       :locationId="locationId"
+      :educationProgram="educationProgram"
     />
   </v-container>
 </template>
@@ -19,6 +20,9 @@
 import { useRouter } from "vue-router";
 import { AESTRoutesConst } from "@/constants/routes/RouteConstants";
 import ManageProgramAndOfferingSummary from "@/components/common/ManageProgramAndOfferingSummary.vue";
+import { ref, onMounted } from "vue";
+import { EducationProgramData } from "@/types";
+import { EducationProgramService } from "@/services/EducationProgramService";
 
 export default {
   components: { ManageProgramAndOfferingSummary },
@@ -38,7 +42,13 @@ export default {
   },
   setup(props: any) {
     const router = useRouter();
+    const educationProgram = ref({} as EducationProgramData);
 
+    const getEducationProgramAndOffering = async () => {
+      educationProgram.value = await EducationProgramService.shared.getEducationProgramForAEST(
+        props.programId,
+      );
+    };
     const goBack = () => {
       router.push({
         name: AESTRoutesConst.INSTITUTION_PROGRAMS,
@@ -47,9 +57,11 @@ export default {
         },
       });
     };
+    onMounted(getEducationProgramAndOffering);
 
     return {
       goBack,
+      educationProgram,
     };
   },
 };
