+++ sources/packages/web/src/views/institution/locations/program-info-request/LocationEditProgramInfoRequest.vue
@@ -32,7 +32,6 @@ import {
   FormIOCustomEvent,
   FormIOCustomEventTypes,
   GetProgramInfoRequestDto,
-  ClientIdType,
 } from "@/types";
 import {
   OFFERING_START_DATE_ERROR,
